Forge fmt (#29)
Add `forge fmt --check` to ci

+++ src/QuarkWallet.sol
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.8.19;
 
-import { CodeJar } from "./CodeJar.sol";
-import { ECDSA } from "@openzeppelin/contracts/utils/cryptography/ECDSA.sol";
+import {CodeJar} from "./CodeJar.sol";
+import {ECDSA} from "@openzeppelin/contracts/utils/cryptography/ECDSA.sol";
 
 contract QuarkWallet {
     error BadSignatory();
@@ -27,10 +27,13 @@ contract QuarkWallet {
     bytes32 internal constant ACTIVE_CALLBACK_SLOT = bytes32(keccak256("org.quark.active-callback"));
 
     /// @dev The EIP-712 typehash for authorizing an operation
-    bytes32 internal constant QUARK_OPERATION_TYPEHASH = keccak256("QuarkOperation(bytes scriptSource,bytes scriptCalldata,uint256 nonce,uint256 expiry,bool allowCallback)");
+    bytes32 internal constant QUARK_OPERATION_TYPEHASH = keccak256(
+        "QuarkOperation(bytes scriptSource,bytes scriptCalldata,uint256 nonce,uint256 expiry,bool allowCallback)"
+    );
 
     /// @dev The EIP-712 typehash for the contract's domain
-    bytes32 internal constant DOMAIN_TYPEHASH = keccak256("EIP712Domain(string name,string version,uint256 chainId,address verifyingContract)");
+    bytes32 internal constant DOMAIN_TYPEHASH =
+        keccak256("EIP712Domain(string name,string version,uint256 chainId,address verifyingContract)");
 
     /// @notice Name of contract, for use in DOMAIN_SEPARATOR
     string public constant name = "Quark Wallet";
@@ -54,9 +57,9 @@ contract QuarkWallet {
         uint256 nonce;
         uint256 expiry;
         bool allowCallback;
-        // requirements
-        // isReplayable
     }
+    // requirements
+    // isReplayable
 
     constructor(address owner_, CodeJar codeJar_) {
         owner = owner_;
@@ -68,7 +71,9 @@ contract QuarkWallet {
          * variable that we are allowed to access with impunity.
          */
         bytes32 slot = OWNER_SLOT;
-        assembly { sstore(slot, owner_) }
+        assembly {
+            sstore(slot, owner_)
+        }
     }
 
     /**
@@ -112,9 +117,7 @@ contract QuarkWallet {
      */
     function DOMAIN_SEPARATOR() public view returns (bytes32) {
         return keccak256(
-            abi.encode(
-                DOMAIN_TYPEHASH, keccak256(bytes(name)), keccak256(bytes(VERSION)), block.chainid, address(this)
-            )
+            abi.encode(DOMAIN_TYPEHASH, keccak256(bytes(name)), keccak256(bytes(VERSION)), block.chainid, address(this))
         );
     }
 
@@ -124,7 +127,9 @@ contract QuarkWallet {
     function getActiveCallback() internal returns (address) {
         bytes32 slot = ACTIVE_CALLBACK_SLOT;
         address callback;
-        assembly { callback := sload(slot) }
+        assembly {
+            callback := sload(slot)
+        }
         return callback;
     }
 
@@ -136,7 +141,9 @@ contract QuarkWallet {
      */
     function setActiveCallback(address callback) internal {
         bytes32 slot = ACTIVE_CALLBACK_SLOT;
-        assembly { sstore(slot, callback) }
+        assembly {
+            sstore(slot, callback)
+        }
     }
 
     /**
@@ -148,16 +155,19 @@ contract QuarkWallet {
      * @param s EIP-712 signature s value
      * @return return value from the executed operation
      */
-    function executeQuarkOperation(
-        QuarkOperation calldata op,
-        uint8 v,
-        bytes32 r,
-        bytes32 s
-    ) public payable returns (bytes memory) {
+    function executeQuarkOperation(QuarkOperation calldata op, uint8 v, bytes32 r, bytes32 s)
+        public
+        payable
+        returns (bytes memory)
+    {
         if (block.timestamp >= op.expiry) revert SignatureExpired();
         if (isSet(op.nonce)) revert InvalidNonce();
 
-        bytes32 structHash = keccak256(abi.encode(QUARK_OPERATION_TYPEHASH, op.scriptSource, op.scriptCalldata, op.nonce, op.expiry, op.allowCallback));
+        bytes32 structHash = keccak256(
+            abi.encode(
+                QUARK_OPERATION_TYPEHASH, op.scriptSource, op.scriptCalldata, op.nonce, op.expiry, op.allowCallback
+            )
+        );
         bytes32 digest = keccak256(abi.encodePacked("\x19\x01", DOMAIN_SEPARATOR(), structHash));
 
         if (isValidSignature(owner, digest, v, r, s)) {
@@ -195,7 +205,11 @@ contract QuarkWallet {
      * @param scriptCalldata The encoded function selector and arguments to call on the transaction script
      * @return return value from the executed operation
      */
-    function executeQuarkOperation(bytes calldata scriptSource, bytes calldata scriptCalldata) public payable returns (bytes memory) {
+    function executeQuarkOperation(bytes calldata scriptSource, bytes calldata scriptCalldata)
+        public
+        payable
+        returns (bytes memory)
+    {
         // XXX authenticate caller
         address scriptAddress = codeJar.saveCode(scriptSource);
         // XXX add support for allowCallback to the direct path
@@ -205,7 +219,10 @@ contract QuarkWallet {
     /**
      * @dev Execute QuarkOperation
      */
-    function executeQuarkOperationInternal(address scriptAddress, bytes memory scriptCalldata, bool allowCallback) internal returns (bytes memory) {
+    function executeQuarkOperationInternal(address scriptAddress, bytes memory scriptCalldata, bool allowCallback)
+        internal
+        returns (bytes memory)
+    {
         uint256 codeLen;
         assembly {
             codeLen := extcodesize(scriptAddress)
@@ -224,10 +241,11 @@ contract QuarkWallet {
         }
 
         bool success;
-        uint returnSize;
-        uint scriptCalldataLen = scriptCalldata.length;
+        uint256 returnSize;
+        uint256 scriptCalldataLen = scriptCalldata.length;
         assembly {
-            success := callcode(gas(), scriptAddress, 0/* value */, add(scriptCalldata, 0x20), scriptCalldataLen, 0x0, 0)
+            success :=
+                callcode(gas(), scriptAddress, 0, /* value */ add(scriptCalldata, 0x20), scriptCalldataLen, 0x0, 0)
             returnSize := returndatasize()
         }
